wip: move //idto/common to //utils
