+++ tools/test.bzl
@@ -1,2 +1,2 @@
 def cc_googletest(name, deps = [], **kwargs):
-  native.cc_test(name=name, deps=deps + ["//idto/common:googletest_main"], **kwargs)
+  native.cc_test(name=name, deps=deps + ["//utils:googletest_main"], **kwargs)
