+++ utils/find_resource.h
