+++ utils/test/profiler_test.cc
@@ -1,4 +1,4 @@
-#include "idto/common/profiler.h"
+#include "utils/profiler.h"
 
 #include <chrono>
 #include <ratio>
